requirements.txt
requests
python-dotenv
termcolor
marshmallow
pipdate
pytest